Split out brick to be stateless component
