+++ components/model.js
@@ -5,6 +5,7 @@ import Task from 'data.task';
 export const resolveImage = (data) => {
   return new Task((reject, resolve) => Image.getSize(data.uri, (width, height) => resolve({
     ...data,
+    //@TODO: Consider consolidating data.uri with dimensions to a image object
     dimensions: {
       width,
       height
